Created conventional Translate and Edit actions
