+++ Apps.OpenAI/Apps.OpenAI.csproj
@@ -9,13 +9,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Blackbird.Applications.Sdk.Common" Version="2.11.0" />
+    <PackageReference Include="Blackbird.Applications.SDK.Blueprints" Version="1.0.0" />
     <PackageReference Include="Blackbird.Applications.SDK.Extensions.FileManagement" Version="1.0.1" />
     <PackageReference Include="Blackbird.Applications.Sdk.Glossaries.Utils" Version="1.0.1" />
     <PackageReference Include="Blackbird.Applications.Sdk.Utils" Version="1.0.25" />
-    <PackageReference Include="Blackbird.Applications.Sdk.Xliff.Utils" Version="1.1.3">
-      <Aliases>XliffContent</Aliases>
-    </PackageReference>
+    <PackageReference Include="Blackbird.Filters" Version="1.1.6" />
     <PackageReference Include="Blackbird.Xliff.Utils" Version="1.1.13" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.17" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
