+++ Apps.OpenAI/Models/Content/ContentProcessingResult.cs
@@ -1,19 +1,17 @@
-﻿extern alias XliffContent;
-
-using Apps.OpenAI.Dtos;
+﻿using Apps.OpenAI.Dtos;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Files;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.Translate;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using XliffContent::Blackbird.Xliff.Utils.Models.Content;
 
 namespace Apps.OpenAI.Models.Content;
-public class ContentProcessingResult
+public class ContentProcessingResult : ITranslateFileOutput
 {
-    public FileReference Content { get; set; }
+    public FileReference File { get; set; }
     public UsageDto Usage { get; set; } = new UsageDto();
 
     [Display("Total segments")]
