+++ Apps.OpenAI/Models/Requests/Content/EditContentRequest.cs
@@ -1,14 +1,14 @@
-﻿using Apps.OpenAI.DataSourceHandlers;
+﻿using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Dictionaries;
 using Blackbird.Applications.Sdk.Common.Files;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Blackbird.Applications.SDK.Blueprints.Handlers;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.Edit;
 
 namespace Apps.OpenAI.Models.Requests.Content;
-public class EditContentRequest
+public class EditContentRequest : IEditFileInput
 {
     public FileReference File { get; set; }
+
+    [Display("Output file handling", Description = "Determine the format of the output file. The default Blackbird behavior is to convert to XLIFF for future steps."), StaticDataSource(typeof(ProcessFileFormatHandler))]
+    public string OutputFileHandling { get; set; }
 }
