+++ Apps.OpenAI/Models/Requests/Content/TranslateContentRequest.cs
@@ -2,14 +2,11 @@
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Dictionaries;
 using Blackbird.Applications.Sdk.Common.Files;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Blackbird.Applications.SDK.Blueprints.Handlers;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.Translate;
 
 namespace Apps.OpenAI.Models.Requests.Content;
-public class TranslateContentRequest
+public class TranslateContentRequest : ITranslateFileInput
 {
     public FileReference File { get; set; }
 
@@ -19,5 +16,8 @@ public class TranslateContentRequest
 
     [Display("Target language")]
     [StaticDataSource(typeof(LocaleDataSourceHandler))]
-    public string? TargetLanguage { get; set; }
+    public string TargetLanguage { get; set; }
+
+    [Display("Output file handling", Description = "Determine the format of the output file. The default Blackbird behavior is to convert to XLIFF for future steps."), StaticDataSource(typeof(ProcessFileFormatHandler))]
+    public string OutputFileHandling { get; set; }
 }
