+++ Tests.OpenAI/Input/contentful.html
@@ -50,7 +50,7 @@
           enjoy the castle’s peaceful grandeur before the crowds arrive.
         </p>
         <p>
-          <strong>Midday: Lunch in Blois</strong> Just a short drive away, the town of Blois offers a taste of
+          <strong>Midday: Lunch in <i>Blois</i></strong> Just a short drive away, the town of Blois offers a taste of
           quintessential Loire Valley charm. Explore its medieval center, winding alleyways, and the regal Château de
           Blois, known for its eclectic mix of architectural styles. For lunch, stop by a local bistro where you can
           savor Loire specialties like goat cheese tartine or rillettes. Pair it with a glass of crisp Sancerre wine,
